final/requirements.txt
# requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.3
langchain_text_splitters>=0.0.1
openai>=1.3.0
supabase>=1.0.3
tqdm>=4.66.1
requests>=2.31.0
pydantic>=2.4.2